third_party/WebKit/LayoutTests/editing/execCommand/4924441.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<p>This tests for a bug where changing the list type of an indented list would create unwanted nesting.  You should see a single ordered list item in an indented list.</p>
<div id="div" contenteditable="true"><br></div>
<div id="log"></div>
<script>
test(function() {
    var div = document.getElementById('div');
    div.focus();
    document.execCommand('InsertText', false, 'foo');
    document.execCommand('InsertUnorderedList');
    document.execCommand('Indent');
    document.execCommand('InsertOrderedList');
    assert_equals(div.innerHTML, '<ul><ol><li>foo</li></ol></ul>');
});
</script>
